% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{add.kernel}
\alias{add.kernel}
\title{Add a Kernel Class to a Model Tree}
\usage{
add.kernel(orig.model.tree, kernel.class.name, kernel, hyper.param.names,
  kernel.additional.params = list(), kernel.grad = NULL)
}
\arguments{
\item{orig.model.tree}{a ModelTree object}

\item{kernel.class.name}{the name of the kernel class (arbitrary, must not
match the class name of any kernel already in the model tree).}

\item{kernel}{either an R function defining a kernel, or a string indicating a built-in kernel type}

\item{hyper.param.names}{a vector containing the names of the kernel hyperparameters}

\item{kernel.additional.params}{a list containing any additional kernel parameters}

\item{kernel.grad}{either an R function returning the gradient of \code{kernel}, or NULL. If \code{kernel}
is a string specifying a built-in kernel, \code{kernel.grad} should be NULL. If \code{kernel} is an
R function and \code{kernel.grad} is NULL, \code{\link[numDeriv]{grad}} will be used to derive a numerical
gradient.}
}
\value{
a ModelTree object
}
\description{
Add a Kernel Class to a Model Tree
}
\examples{
mt <- create.model.tree()
mt <- add.kernel(mt, "SE", "squaredExponential", c("l"))
}

