% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{add.polynomial.kernels}
\alias{add.polynomial.kernels}
\title{Add Polynomial Kernel Classes to a Model Tree}
\usage{
add.polynomial.kernels(mt, degrees = c(1, 2), homogeneous = TRUE,
  poly = TRUE, generalised = TRUE)
}
\arguments{
\item{mt}{a ModelTree object}

\item{degrees}{a vector containing the degrees to add}

\item{homogeneous}{boolean indicating whether to add homogeneous polynomial kernels}

\item{poly}{boolean indicating whether to add non-homogeneous polynomial kernels}

\item{generalised}{boolean indicating whether to add generealised polynomial kernels}
}
\value{
a ModelTree object
}
\description{
Add Polynomial Kernel Classes to a Model Tree
}
\examples{
mt <- create.model.tree()
mt <- add.polynomial.kernels(mt, degrees=c(2,3))
}

