% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{as.character.ModelTree}
\alias{as.character.ModelTree}
\alias{print.ModelTree}
\title{Convert a Model Tree to a String}
\usage{
\method{as.character}{ModelTree}(model.tree, show.node.ids = FALSE)

\method{print}{ModelTree}(model.tree, show.node.ids = FALSE)
}
\arguments{
\item{model.tree}{a modelTree object}

\item{show.node.ids}{boolean indicating whether the string should be annotated with the associated node numbers.}
}
\value{
a string representing a modelTree object
}
\description{
Return or print a string representation of the combination of kernels represented by a given model tree
}
\examples{
mt <- create.model.tree.builtin()
mt <- insert.kernel.instance(mt, 1, "SE", NULL)
mt <- insert.kernel.instance(mt, 1, "SE", "+")

as.character(mt) #"( SE.1 + SE.2 )"

as.character(mt, show.node.ids=TRUE) #"[3: ( [1: SE.1 :1] + [2: SE.2 :2] ) :3]"

print(mt) #"( SE.1 + SE.2 )"

print(mt, show.node.ids=TRUE) #"[3: ( [1: SE.1 :1] + [2: SE.2 :2] ) :3]"
}

