% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{bayesian.information.criterion}
\alias{bayesian.information.criterion}
\title{Calculate BIC of a GP}
\usage{
bayesian.information.criterion(gp.obj)
}
\arguments{
\item{gp.obj}{A trained gaussianProcess object}
}
\value{
The BIC value for the input GP object.
}
\description{
Calculate BIC of a GP
}
\examples{
x <- rnorm(50)
y <- sin(1/(x^2 + 0.15))
mt <- create.model.tree.builtin()
mt <- insert.kernel.instance(mt, 1, "SE", NULL, hyper.params=c(l=1))
gp <- create.gaussian.process(x, y, mt)
gp$fit.hyperparams(NA)
}
\seealso{
\code{\link{bic.model.search}} and \code{\link{gaussianProcess}}
}

