% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{callKernelByString}
\alias{callKernelByString}
\title{Select Built-in C++ Kernels by Name}
\usage{
callKernelByString(kernelName, a, b, hyperParams, additionalParams)
}
\arguments{
\item{kernelName}{the kernel's name (as a string)}

\item{a}{the first data point}

\item{b}{the second data point}

\item{hyperParams}{the kernel's hyperparameters as a named numeric vector}

\item{additionalParams}{a list of any additional parameters}
}
\description{
Built in kernels include:
\itemize{
  \item squaredExponential
  \item rationalQuadratic
  \item periodic
  \item constant
  \item generalisedLinear
  \item oneDLinear
  \item changepoint
  \item randomForest
  \item neuralNetwork
  \item generalisedPolynomial
  \item polynomial
  \item homogeneousPolynomial
}
}

