% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{clone.env}
\alias{clone.env}
\title{Clone an Environment}
\usage{
clone.env(orig.env)
}
\arguments{
\item{orig.env}{the environment to clone}
}
\value{
a new environment whose contents match \code{orig.env}
}
\description{
Clone an Environment
}
\examples{
mt <- create.model.tree.builtin()
mt2 <- clone.env(mt)
}
\seealso{
\code{\link{clone.ModelTree}}
}

