% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{clone.ModelTree}
\alias{clone.ModelTree}
\title{Clone a Model Tree}
\usage{
clone.ModelTree(orig.model.tree, invalidate.cache = TRUE)
}
\arguments{
\item{orig.model.tree}{the ModelTree to clone}

\item{invalidate.cache}{boolean indicating whether the cache should be cleared.
If you're not sure whether you need to clear the cache or not, you probably do.}
}
\value{
a new Model Tree whose contents match \code{orig.model.tree}
}
\description{
Clone a Model Tree
}
\examples{
mt <- create.model.tree.builtin()
mt2 <- clone.ModelTree(mt)
}

