% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create.ard.kernel}
\alias{create.ard.kernel}
\title{Create an ARD kernel object}
\usage{
create.ard.kernel(dimensions, inverse = T)
}
\arguments{
\item{dimensions}{}
}
\value{
a kernel object
}
\description{
Create an ARD kernel object
}

