% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{create.gaussian.process}
\alias{create.gaussian.process}
\title{Create a gaussianProcess object}
\usage{
create.gaussian.process(x, y, kernel, cache = NULL)
}
\arguments{
\item{x}{A matrix or data frame of predictor variables}

\item{y}{A numeric vector of response variables}

\item{kernel}{a Kernel object specifying the kernel for the Gaussian process}

\item{cache}{a Cache object. If NULL, a cache is created}
}
\value{
An untrained gaussianProcess object
}
\description{
This function creates a gaussianProcess object along with associated functions for fitting hyperparameters.
}
\details{
Intend to rework this to fit the R S3 framework - it currently encloses all its methods to allow caching in
optimx calls, but with a bit more experience I think this isn't actually necessary.
}
\examples{
x <- rnorm(50)
y <- sin(1/(x^2 + 0.15))
mt <- create.model.tree.builtin()
mt <- insert.kernel.instance(mt, 1, "SE", NULL, hyper.params=c(l=1))
gp <- create.gaussian.process(x, y, mt)
gp$fit.hyperparams(NA)

}

