% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create.kernel.object.from.model.tree}
\alias{create.kernel.object.from.model.tree}
\title{Create Kernel object from model tree}
\usage{
create.kernel.object.from.model.tree(model_tree)
}
\arguments{
\item{model_tree}{}
}
\value{
a kernel object
}
\description{
Create Kernel object from model tree
}

