% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create.kernel.object}
\alias{create.kernel.object}
\title{Create a kernel object}
\usage{
create.kernel.object(kernel, grad_function = NULL, hess_function = NULL,
  hyperparam_names = character(0), additional_params = list())
}
\arguments{
\item{grad_function}{if \code{kernel_function} is a function, then this
should be a function returning the grad w.r.t. the hyperparameters,
otherwise this should be null}

\item{hyperparam_names}{A character vector of the names of the hyperparameters}

\item{additional_params}{A list of additional parameters}

\item{kernel_function}{the kernel function (either a string, a function,
or a ModelTree)}
}
\value{
A Kernel object
}
\description{
Kernel objects provide a unifying wrapper for the various things which
can act as kernels (built-in kernels (represented by strings), model trees,
and arbitrary R functions).
}

