% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{create.model.tree.builtin}
\alias{create.model.tree.builtin}
\title{Create a Model Tree Containing All Built-in Kernels}
\usage{
create.model.tree.builtin()
}
\value{
A modelTree object
}
\description{
Returns a model tree instance populated with a basis set of all of the built in kernels.
}
\examples{
mt <- create.model.tree.builtin()
}

