% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{create.model.tree.from.list}
\alias{create.model.tree.from.list}
\title{Create a Model Tree from a list of kernel objects}
\usage{
create.model.tree.from.list(kernel_list)
}
\arguments{
\item{kernel_list}{a named list of kernel objects}
}
\value{
A modelTree object
}
\description{
Returns a model tree instance populated with a basis set of kernels.
}

