% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{create.model.tree}
\alias{create.model.tree}
\title{Create an Empty Model Tree}
\usage{
create.model.tree()
}
\value{
a modelTree object
}
\description{
Creates an instance of a modelTree object with no associated kernels.
}
\examples{
mt <- create.model.tree()
}

