% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create.numeric.grad}
\alias{create.numeric.grad}
\title{Create a numerical gradient function for a kernel using the \code{\link[numDeriv]{grad}} function.}
\usage{
create.numeric.grad(k, method = "Richardson", method.args = NULL,
  additional.params = list())
}
\arguments{
\item{k}{kernel function}

\item{method}{method for creating numeric grad, see \code{\link[numDeriv]{grad}}}

\item{method.args}{method arguments for creating numeric grad, see \code{\link[numDeriv]{grad}}}

\item{additional.params}{additional parameters for the kernel function \code{k}}
}
\value{
A function which gives the numeric grad of \code{k} w.r.t. the kernel hyperparameters.
}
\description{
Create a numerical gradient function for a kernel using the \code{\link[numDeriv]{grad}} function.
}

