% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create.numeric.hessian.from.grad}
\alias{create.numeric.hessian.from.grad}
\title{Create a numerical Hessian function from a kernel grad using the \code{\link[numDeriv]{jacobian}} function.}
\usage{
create.numeric.hessian.from.grad(k.grad, method = "Richardson",
  method.args = NULL, additional.params = list())
}
\arguments{
\item{k.grad}{function giving the gradient of the kernel}

\item{method}{method for creating numeric Jacobian, see \code{\link[numDeriv]{jacobian}}}

\item{method.args}{method arguments for creating numeric Jacobian, see \code{\link[numDeriv]{jacobian}}}

\item{additional.params}{additional parameters for the kernel grad function \code{k}}
}
\value{
A function which gives the numeric Hessian for \code{k.grad} w.r.t. the kernel hyperparameters.
}
\description{
Create a numerical Hessian function from a kernel grad using the \code{\link[numDeriv]{jacobian}} function.
}

