% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{create.numeric.hessian}
\alias{create.numeric.hessian}
\title{Create a numerical Hessian function for a kernel using the \code{\link[numDeriv]{hessian}} function.}
\usage{
create.numeric.hessian(k, method = "Richardson", method.args = NULL,
  additional.params = list())
}
\arguments{
\item{k}{kernel function}

\item{method}{method for creating numeric Hessian, see \code{\link[numDeriv]{hessian}}}

\item{method.args}{method arguments for creating numeric Hessian, see \code{\link[numDeriv]{hessian}}}

\item{additional.params}{additional parameters for the kernel function \code{k}}
}
\value{
A function which gives the numeric Hessian of \code{k} w.r.t. the kernel hyperparameters.
}
\description{
Create a numerical Hessian function for a kernel using the \code{\link[numDeriv]{hessian}} function.
}

