% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfTreeTools.R
\name{create.rf.additional.params}
\alias{create.rf.additional.params}
\title{Create Additional Params List for Random Forest Kernel}
\usage{
create.rf.additional.params(rf)
}
\arguments{
\item{rf}{a randomForest object}
}
\value{
a list containing the additional params for \code{\link{randomForestKernel}}.
}
\description{
Takes a \code{\link[randomForest]{randomForest}} object and extracts the
components required by \code{\link{randomForestKernel}} and
\code{\link{randomForestKernelGrad}}
}

