% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{delete.node}
\alias{delete.node}
\title{Delete a Node in a Model Tree}
\usage{
delete.node(model.tree, node)
}
\arguments{
\item{model.tree}{a ModelTree object}

\item{node}{the node to delete (must not be the root node)}
}
\value{
a clone of \code{model.tree} with \code{node} deleted.
}
\description{
Delete a Node in a Model Tree
}

