% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{eval.cov.mat.grad.node}
\alias{eval.cov.mat.grad.node}
\title{Evaluate Covariance Matrix Grad of a Model Tree Node}
\usage{
eval.cov.mat.grad.node(model.tree, inst.cov.mats, inst.cov.mat.grads, node)
}
\arguments{
\item{model.tree}{An object of class modelTree}

\item{inst.cov.mats}{a named list of the covariance matrices of the kernel instances in the model tree}

\item{inst.cov.mat.grads}{The covariance matrix grads of the kernel instances in the model tree}

\item{node}{The node to evaluate}
}
\value{
A covariance matrix grad (as an array)
}
\description{
This function recursively evaluates a model tree at the given node, returning
 the grad of the covariance matrix at that node
}

