% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{eval.cov.mat.node}
\alias{eval.cov.mat.node}
\title{Evaluate Covariance Matrix of a Model Tree Node}
\usage{
eval.cov.mat.node(model.tree, inst.cov.mats, node)
}
\arguments{
\item{model.tree}{an object of class modelTree}

\item{inst.cov.mats}{a named list of the covariance matrices of the kernel instances in the model tree}

\item{node}{the node to evaluate}
}
\value{
A covariance matrix
}
\description{
This function recursively evaluates a model tree at the given node,
 returning the covariance matrix for that node
}

