% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{eval.node.kstar.mat}
\alias{eval.node.kstar.mat}
\title{K* matrix of a model tree node}
\usage{
eval.node.kstar.mat(model.tree, inst.kstar.mats, node)
}
\arguments{
\item{model.tree}{a model tree}

\item{inst.kstar.mats}{a named list of the K* matrices of the kernel instances in the model tree}

\item{node}{the node to evaluate}
}
\value{
a numeric matrix
}
\description{
Recursively evaluates the K* matrix of a model tree from a given node.
}

