% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{fit.hyperparams}
\alias{fit.hyperparams}
\title{Fit hyperparameters for a GP using maximum likelihood optimisation}
\usage{
fit.hyperparams(gp, sigma.n.init = NA, hyper.params.init = NA,
  random.init.gridsize = 50, resample.top.inits = TRUE,
  resample.from = NULL, num.resamples = NULL, additional.optimx.runs = 1,
  additional.par.perc.diff = 0.1, random.init.scale = 1,
  abs.min.sigma.n = 0.001, abs.min.hyper.params = 0,
  optimx.method = "L-BFGS-B", max.iterations = 10000,
  optimx.starttests = FALSE, optimx.trace = 0, verbose = FALSE)
}
\arguments{
\item{gp}{a GaussianProcess object}

\item{sigma.n.init}{an initial value for sigma.n, or NA for random
initialisation}

\item{hyper.params.init}{a vector of initial values for the kernel
hyperparameters, or NA for random initialisation. Partial random starts
can be specified by setting only some entries in the vector to be NA, or
a complete random start can be specified by passing a single NA value.}

\item{random.init.gridsize}{the number of random initialisation points to
test}

\item{resample.top.inits}{boolean indicating whether to do a second random
search based on the (smoothed) distribution of the top results in the first
pass}

\item{resample.from}{number of points to build the distribution from in the
second random search}

\item{num.resamples}{number of samples for the second search}

\item{additional.optimx.runs}{the number of additional start points to attempt
optimisation from}

\item{additional.par.perc.diff}{the percentage difference required between
any two starting parameter vectors}

\item{random.init.scale}{a scaling factor for the random generation of
initialisation points}

\item{abs.min.sigma.n}{minimum value of sigma.n}

\item{abs.min.hyper.params}{a vector of minimum values of the hyper params}

\item{optimx.method}{the optimx method to use for optimisation}

\item{max.iterations}{max iterations of each optimx run}

\item{optimx.starttests}{run optimx start tests}

\item{optimx.trace}{optimx trace level}

\item{verbose}{verbose output to screen}
}
\value{
a trained GaussianProcess object
}
\description{
Fit hyperparameters for a GP using maximum likelihood optimisation
}

