% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\docType{package}
\name{gaussianProcess-package}
\alias{gaussianProcess}
\alias{gaussianProcess-package}
\title{Utilities for training Gaussian processes.}
\description{
gaussianProcess provides functionality to define kernels (either
stand-alone or as a combination of basic kernels) and also
to automatically perform model selection from a basic set of kernels
to select the optimum kernel for modelling the input data, using the
Bayesian Information Criterion as a search criteria.
}

