% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussianProcess.R
\name{gaussianProcess}
\alias{gaussianProcess}
\title{Create a gaussianProcess object with a given mean function and covariance
kernel, and data}
\usage{
gaussianProcess(X, y, meanFunc = zeroFunction, kernel = rbf, noiseVar = 1,
  scale = 1, order = NULL, amplitude = 1)
}
\arguments{
\item{X}{An n x d matrix of covariates for observed data}

\item{y}{An n dimensional vector of outputs for observed data}

\item{meanFunc}{The mean function of the process, defaults to 0}

\item{kernel}{The covariance kernel of the process, defaults to rbf}

\item{noiseVar}{The variance of the noise around the function}

\item{scale}{The scale of the kernel, defaults to 1}

\item{order}{The order of the kernel, defaults to NULL}

\item{amplitude}{A parameter which is multiplied by the kernel, defaults to 1}
}
\value{
A gaussianProcess with these options
}
\description{
Create a gaussianProcess object with a given mean function and covariance
kernel, and data
}

