% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{get.alpha}
\alias{get.alpha}
\title{Get alpha for a GP}
\usage{
get.alpha(kernel, training.points, training.point.values, sigma.n, hyper.params,
  cache)
}
\arguments{
\item{gp}{a GaussianProcess object}
}
\value{
a vector
}
\description{
Internal shenanigans. This documentation only exists so I can tell Roxygen
what to import for it.
}

