% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{get.cov.mat.chol}
\alias{get.cov.mat.chol}
\title{Get the Cholesky decomposition of a covariance matrix}
\usage{
get.cov.mat.chol(kernel, x, sigma.n, hyper.params, cache)
}
\value{
the Cholesky decomposition of a covariance matrix
}
\description{
Get the Cholesky decomposition of a covariance matrix
}

