% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.covariance.matrix.grad.kernel}
\alias{get.covariance.matrix.grad.kernel}
\title{Get the Covariance Matrix Grad for a Kernel}
\usage{
get.covariance.matrix.grad.kernel(x, kernel.grad, sigma.n, hyper.params,
  additional.params)
}
\arguments{
\item{x}{A matrix of data to find the covariance matrix of}

\item{kernel.grad}{A function returning the grad of a kernel (as a named vector)}

\item{sigma.n}{The standard deviation of the noise to add to the kernel - used to regularise the resulting covariance matrix}

\item{hyper.params}{The hyperparameters of \code{kernel.grad}}

\item{additional.params}{Any additional parameters of \code{kernel.grad}}
}
\value{
A covariance matrix grad (as an array)
}
\description{
Get the Covariance Matrix Grad for a Kernel
}

