% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.covariance.matrix.grad.model.tree}
\alias{get.covariance.matrix.grad.model.tree}
\title{Covariance Matrix Grad of a Model Tree}
\usage{
get.covariance.matrix.grad.model.tree(x, model.tree, sigma.n)
}
\arguments{
\item{x}{A numeric matrix}

\item{model.tree}{A modelTree object}

\item{sigma.n}{The standard deviation of the noise to add to the kernel - used to regularise the resulting covariance matrix}
}
\value{
An array containing the covariance matrix grad.
}
\description{
Calculates the covariance matrix grad for a modelTree object
}
\examples{
mt <- create.model.tree.builtin()
mt <- insert.kernel.instance(mt, 1, "SE", NULL, hyper.params=c(l=1))

x <- rnorm(50)

cov.mat <- get.covariance.matrix.model.tree(x, mt, sigma.n=0.1)
}

