% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.covariance.matrix.kernel}
\alias{get.covariance.matrix.kernel}
\title{Get the Covariance Matrix for a Kernel}
\usage{
get.covariance.matrix.kernel(x, k, sigma.n, hyper.params, additional.params)
}
\arguments{
\item{x}{A matrix of data to find the covariance matrix of}

\item{k}{A kernel function}

\item{sigma.n}{The standard deviation of the noise to add to the kernel - used to regularise the resulting covariance matrix}

\item{hyper.params}{The hyperparameters of \code{k}}

\item{additional.params}{Any additional parameters of \code{k}}
}
\value{
A covariance matrix
}
\description{
Get the Covariance Matrix for a Kernel
}

