% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.inst.cov.mats}
\alias{get.inst.cov.mats}
\title{Get Kernel Instance Covariance Matrices or Grads}
\usage{
get.inst.cov.mats(x, model.tree, return.grad = FALSE)
}
\arguments{
\item{x}{A numeric matrix}

\item{model.tree}{A modelTree object}

\item{return.grad}{Whether to return the covariance matrices or the associated grads}
}
\value{
A named list containing covariance matrices (or grad arrays), with one entry for each kernel instance.
}
\description{
Calculates the covariance matrices (or associated grads) for the kernel instances appearing in a model tree
}

