% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.kstar.mat.kernel}
\alias{get.kstar.mat.kernel}
\title{Get predictive K* matrix}
\usage{
get.kstar.mat.kernel(data.to.predict, training.data, k, hyper.params,
  additional.params)
}
\arguments{
\item{data.to.predict}{numeric matrix of data to predict}

\item{training.data}{the data used to train the Gaussian process}

\item{k}{the kernel function}

\item{hyper.params}{the kernel's hyperparameters}

\item{additional.params}{the kernel's additional parameters}
}
\value{
a numeric matrix
}
\description{
This function returns the K* matrix of covariances between the training data and the data to predict.
}

