% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.kstar.mat.model.tree}
\alias{get.kstar.mat.model.tree}
\title{K* Matrix of a Model Tree}
\usage{
get.kstar.mat.model.tree(data.to.predict, training.data, model.tree)
}
\arguments{
\item{data.to.predict}{a numeric matrix}

\item{training.data}{a numeric matrix}

\item{model.tree}{a modelTree object}
}
\value{
a numeric matrix
}
\description{
Returns the K* matrix between training data and data to predict for a given modelTree object.
}

