% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{get.marginal.likelihood.hessian}
\alias{get.marginal.likelihood.hessian}
\title{Get the Hessian of the Log Marginal Likelihood of a GP}
\usage{
get.marginal.likelihood.hessian(gp, sigma.n, hyper.params)
}
\arguments{
\item{gp}{a GaussianProcess object}

\item{sigma.n}{a numeric value for sigma.n}

\item{hyper.params}{a numeric vector of hyperparameters}
}
\value{

}
\description{
Returns the Hessian matrix of the log marginal likelihood of a GP w.r.t. the kernel's
hyperparameters.
}

