% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{get.marginal.likelihood}
\alias{get.marginal.likelihood}
\title{Get Log Marginal Likelihood of a GP}
\usage{
get.marginal.likelihood(gp, sigma.n, hyper.params)
}
\arguments{
\item{gp}{a GaussianProcess object}

\item{sigma.n}{a numeric value for sigma.n}

\item{hyper.params}{a numeric vector of hyperparameters}
}
\value{
the log marginal likelihood of the GP at the given hyperparameters
}
\description{
Get Log Marginal Likelihood of a GP
}

