% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get.normal.prior}
\alias{get.normal.prior}
\title{Return an object representing a normal prior over hyperparameters}
\usage{
get.normal.prior(sigma)
}
\arguments{
\item{sigma}{the standard deviation of the prior}
}
\value{
a HyperParamPrior object
}
\description{
Return an object representing a normal prior over hyperparameters
}

