% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{get.uniform.prior}
\alias{get.uniform.prior}
\title{Return an object representing a uniform prior over hyperparameters}
\usage{
get.uniform.prior()
}
\value{
a HyperParamPrior object
}
\description{
Return an object representing a uniform prior over hyperparameters
}

