% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_covariance_matrix.ModelTree}
\alias{get_covariance_matrix.ModelTree}
\title{Get the Covariance Matrix for a Kernel ModelTree}
\usage{
\method{get_covariance_matrix}{ModelTree}(kernel, x, sigma.n, hyper.params,
  additional.params = list())
}
\arguments{
\item{kernel}{A ModelTree object}

\item{x}{A matrix of data to find the covariance matrix of}

\item{sigma.n}{The standard deviation of the noise to add to the kernel - used to regularise the resulting covariance matrix}

\item{hyper.params}{The hyperparameters of \code{kernel}}

\item{additional.params}{an empty list (not used for ModelTree objects)}
}
\value{
A covariance matrix
}
\description{
Get the Covariance Matrix for a Kernel ModelTree
}

