% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_covariance_matrix_hess.character}
\alias{get_covariance_matrix_hess.character}
\title{Get the Covariance Matrix Hessian for a built-in kernel}
\usage{
\method{get_covariance_matrix_hess}{character}(kernel, kernel.hess = NULL, x,
  sigma.n, hyper.params, additional.params, cache = NULL)
}
\arguments{
\item{kernel}{A string}

\item{kernel.hess}{Not used for built in kernels}

\item{x}{A matrix of data to find the covariance matrix Hessian of}

\item{sigma.n}{The standard deviation of the noise to add to the kernel - used to regularise the resulting covariance matrix}

\item{hyper.params}{The hyperparameters of \code{kernel}}

\item{additional.params}{Any additional parameters of \code{kernel} (internal use only)}
}
\value{
A covariance matrix
}
\description{
Get the Covariance Matrix Hessian for a built-in kernel
}

