% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_covariance_matrix_hess.function}
\alias{get_covariance_matrix_hess.function}
\title{Get the Covariance Matrix Hessian for a kernel function}
\usage{
\method{get_covariance_matrix_hess}{function}(kernel, kernel.hess, x, sigma.n,
  hyper.params, additional.params, cache = cache)
}
\arguments{
\item{kernel}{A kernel function}

\item{kernel.hess}{A Hessian function for \code{kernel}}

\item{x}{A matrix of data to find the covariance matrix grad of}

\item{sigma.n}{The standard deviation of the noise to add to the kernel - used to regularise the resulting covariance matrix}

\item{hyper.params}{The hyperparameters of \code{kernel}}

\item{additional.params}{Any additional parameters of \code{kernel} (internal use only)}
}
\value{
A covariance matrix
}
\description{
Get the Covariance Matrix Hessian for a kernel function
}

