% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_kstar_matrix.function}
\alias{get_kstar_matrix.function}
\title{Get predictive K* matrix}
\usage{
\method{get_kstar_matrix}{function}(kernel, data.to.predict, training.data,
  hyper.params, additional.params)
}
\arguments{
\item{data.to.predict}{numeric matrix of data to predict}

\item{training.data}{the data used to train the Gaussian process}

\item{hyper.params}{the kernel's hyperparameters}

\item{additional.params}{the kernel's additional parameters}

\item{k}{the kernel function}
}
\value{
a numeric matrix
}
\description{
This function returns the K* matrix of covariances between the
training data and the data to predict for an R function
}

