% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_kstar_matrix.Kernel}
\alias{get_kstar_matrix.Kernel}
\title{Get predictive K* matrix}
\usage{
\method{get_kstar_matrix}{Kernel}(kernel, data.to.predict, training.data,
  hyper.params)
}
\arguments{
\item{kernel}{the Kernel object}

\item{data.to.predict}{numeric matrix of data to predict}

\item{training.data}{the data used to train the Gaussian process}

\item{hyper.params}{the kernel's hyperparameters}
}
\value{
a numeric matrix
}
\description{
This function returns the K* matrix of covariances between the
training data and the data to predict for a Kernel object
}

