% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_kstar_matrix.ModelTree}
\alias{get_kstar_matrix.ModelTree}
\title{K* Matrix of a Model Tree}
\usage{
\method{get_kstar_matrix}{ModelTree}(kernel, data.to.predict, training.data,
  hyper.params, additional.params = list())
}
\arguments{
\item{kernel}{a ModelTree object}

\item{data.to.predict}{numeric matrix of data to predict}

\item{training.data}{the data used to train the Gaussian process}

\item{hyper.params}{the ModelTree's hyperparameters}

\item{additional.params}{an empty list (not used for ModelTree objects)}
}
\value{
a numeric matrix
}
\description{
Returns the K* matrix between training data and data to predict for a given modelTree object.
}

