% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get_kstar_matrix}
\alias{get_kstar_matrix}
\title{Get predictive K* matrix}
\usage{
get_kstar_matrix(kernel, data.to.predict, training.data, hyper.params, ...)
}
\arguments{
\item{kernel}{the kernel function}

\item{data.to.predict}{numeric matrix of data to predict}

\item{training.data}{the data used to train the Gaussian process}

\item{hyper.params}{the kernel's hyperparameters}
}
\value{
a numeric matrix
}
\description{
This function returns the K* matrix of covariances between the
training data and the data to predict
}

