% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{gp_lml_deriv}
\alias{gp_lml_deriv}
\title{Compute the derivative of the log marginal likelihood}
\usage{
gp_lml_deriv(hyper.params, X, y, kernel.type, order = 5/2, noise.var = 1)
}
\arguments{
\item{hyper.params}{Hyper parameters as a vector (for rbf c(amplitude, scales))}

\item{X}{An n x d matrix of covariates for observed data}

\item{y}{An n dimensional vector of outputs for observed data}

\item{kernel.type}{The kernel function to use}

\item{order}{The order of the kernel, defaults to 5/2}

\item{noise.var}{The variance of the noise around the function}
}
\value{
The gradient of the log marginal likelihood at \code{hyper.params}
}
\description{
Compute the derivative of the log marginal likelihood
}

