% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{gp_lml_deriv_helper}
\alias{gp_lml_deriv_helper}
\title{Compute the derivative of the log marginal likelihood (helper function)}
\usage{
gp_lml_deriv_helper(hyper.params, diff.mat, y, K, K.inv, alpha, kernel.deriv,
  noise.var)
}
\arguments{
\item{hyper.params}{Hyper parameters as a vector (for rbf c(amplitude, scales))}

\item{diff.mat}{The result of \code{pdiff(X,X)} for input data \code{X}}

\item{y}{The target vector}

\item{K}{The kernel matrix of \code{X}}

\item{K.inv}{The inverse of \code{K} (precomputed)}

\item{alpha}{An n dimensional vector, alpha = K^-1 y}

\item{kernel.deriv}{The derivative function for the kernel}

\item{noise.var}{The variance of the noise around the function}
}
\value{
The gradient of the log marginal likelihood at \code{hyper.params}
}
\description{
Compute the derivative of the log marginal likelihood (helper function)
}

