% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{gp_log_marg_like}
\alias{gp_log_marg_like}
\title{Compute the marginal log likelihood for a GP with given data, kernel, etc.}
\usage{
gp_log_marg_like(hyper.params, X, y, kernel.type = rbf, order = 5/2,
  noise.var = 1)
}
\arguments{
\item{hyper.params}{Hyper parameters as a vector (for rbf c(amplitude, scales))}

\item{X}{An n x d matrix of covariates for observed data}

\item{y}{An n dimensional vector of outputs for observed data}

\item{kernel.type}{The kernel function to use}

\item{order}{The order of the kernel, defaults to 5/2}

\item{noise.var}{The variance of the noise around the function}
}
\value{
Log marginal likelihood
}
\description{
Compute the marginal log likelihood for a GP with given data, kernel, etc.
}

