% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{gp_log_marg_like_helper}
\alias{gp_log_marg_like_helper}
\title{Compute the log marginal likelihood from parameters of the GP}
\usage{
gp_log_marg_like_helper(y, K, alpha)
}
\arguments{
\item{y}{An n dimensional vector of outputs for observed data}

\item{K}{The n x n covariance matrix on the seen data}

\item{alpha}{An n dimensional vector, alpha = K^-1 y}
}
\value{
Log marginal likelihood
}
\description{
Compute the log marginal likelihood from parameters of the GP
}

