% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{insert.kernel.instance}
\alias{insert.kernel.instance}
\title{Insert Kernel Instance into Model Tree}
\usage{
insert.kernel.instance(orig.model.tree, node, kernel.class.name, operation.id,
  hyper.params = NULL)
}
\arguments{
\item{orig.model.tree}{a ModelTree object}

\item{node}{the node at which to insert the kernel instance}

\item{kernel.class.name}{a string indicating the class of kernel to insert}

\item{operation.id}{a string indicating how to insert the kernel at \code{node}.
Must be "*" or "+", or NULL if the model tree is currently empty.}

\item{hyper.params}{the hyper parameters of the inserted kernel. Can be NULL to
fit hyper parameters at a later time.}
}
\value{
a ModelTree object
}
\description{
Insert Kernel Instance into Model Tree
}
\examples{
mt <- create.model.tree.builtin()

mt <- insert.kernel.instance(mt, 1, "SE", NULL)
print(mt) # "SE.1"

mt <- insert.kernel.instance(mt, 1, "SE", "*")
print(mt) # "( SE.1 * SE.2 )"

mt <- insert.kernel.instance(mt, 3, "RQ", "+")
print(mt) # "( ( SE.1 * SE.2 ) + RQ.1 )"

}

