% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelUtils.R
\name{list_built_in_kernels}
\alias{list_built_in_kernels}
\title{Return a list containing built in kernels.}
\usage{
list_built_in_kernels()
}
\value{
A list of built in kernel objects
}
\description{
Not all built in kernels are included - some are data-dependent (RF kernel)
and others require additional things to be specified like degree or the
dimensions of the data they act on.
}
\examples{
built_in_kernels <- list_built_in_kernels()
built_in_kernels
}

