% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelUtils.R
\name{list_built_in_kernels_scaled}
\alias{list_built_in_kernels_scaled}
\title{Return a list of built-in kernels scaled with a constant kernel}
\usage{
list_built_in_kernels_scaled()
}
\value{
a list of Kernel objects
}
\description{
Return a list of built-in kernels scaled with a constant kernel
}

