% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelUtils.R
\name{list_polynomial_kernels}
\alias{list_polynomial_kernels}
\title{Create a list of polynomial kernels}
\usage{
list_polynomial_kernels(degrees = c(1, 2), homogeneous = TRUE,
  poly = TRUE, generalised = TRUE)
}
\arguments{
\item{degrees}{a vector containing the degrees to include}

\item{homogeneous}{boolean indicating whether to add homogeneous polynomial kernels}

\item{poly}{boolean indicating whether to add non-homogeneous polynomial kernels}

\item{generalised}{boolean indicating whether to add generalised polynomial kernels}
}
\value{
a list of Kernel objects
}
\description{
Create a list of polynomial kernels
}
\examples{
mt <- create.model.tree()
mt <- add.polynomial.kernels(mt, degrees=c(2,3))
}

