% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{log_marginal_like.gaussianProcess}
\alias{log_marginal_like.gaussianProcess}
\title{Compute the log marginal likelihood for a GP}
\usage{
\method{log_marginal_like}{gaussianProcess}(gp)
}
\arguments{
\item{gp}{A gaussianProcess object}
}
\value{
Log marginal likelihood
}
\description{
Compute the log marginal likelihood for a GP
}

