% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{loocv.gaussianProcess}
\alias{loocv.gaussianProcess}
\title{Compute the leave one out CV estimate (Rasmussen & WIllaims pg 117)}
\usage{
\method{loocv}{gaussianProcess}(gp)
}
\arguments{
\item{gp}{The gaussianProcess}
}
\description{
Compute the leave one out CV estimate (Rasmussen & WIllaims pg 117)
}

