% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{matern}
\alias{matern}
\title{Compute the Matern kernel between two matrices}
\usage{
matern(X, Y, scale = 1, order = 5/2, amplitude = 1)
}
\arguments{
\item{X}{An n x d matrix}

\item{Y}{An n x d matrix}

\item{scale}{The scale parameter, defaults to 1}

\item{order}{whether to compute a 3/2 or 5/2 kernel, defaults to 5/2}

\item{amplitude}{A parameter which is multiplied by the kernel, defaults to 1}
}
\value{
An n x n kernel matrix formed from \code{X} and \code{Y}
}
\description{
Compute the Matern kernel between two matrices
}

