% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{model.search}
\alias{model.search}
\title{Perform Model Selection Using BIC}
\usage{
model.search(x, y, base.model.tree, plot.gp = FALSE, reset.params = FALSE,
  max.new.nodes = 3, revisit.kernels = TRUE,
  scoring.function = bayesian.information.criterion,
  return.all.models = FALSE, alternate.scoring.functions = list(), ...)
}
\arguments{
\item{x}{A matrix or data frame of predictors}

\item{y}{A numeric vector of responses}

\item{base.model.tree}{The model tree at which the search starts}

\item{plot.gp}{Whether to plot the gaussian processes encountered during the search. If ncol(x) > 1 this parameter is ignored and no plots are created.}

\item{reset.params}{By default the search starts with the optimum hyperparameter values found in the previous step in the search (with new hyperparameters fitted from random start points). Setting this to TRUE causes all hyperparameters to be randomly set at each step.}

\item{max.new.nodes}{The number of kernel instances to add to the base model.}

\item{revisit.kernels}{Whether to revisit previously-assessed kernels when deleting nodes from the current best candidate.}

\item{scoring.function}{The scoring function for the GPs in the search. Must take a trained gaussianProcess object as its only parameter, and return a scalar score. If \code{scoring.function} returns a numeric vector only the first entry will be used.}

\item{...}{Additional parameters to be passed to gp.obj$fit.hyperparameters (see \code{\link{gaussianProcess}})}
}
\value{
The trained gaussianProcess object with the best BIC.
}
\description{
Perform Model Selection Using BIC
}
\examples{
x <- rnorm(50)
y <- sin(1/(x^2 + 0.15))
mt <- create.model.tree.builtin()
gp <- bic.model.search(x, y, mt)
}

