% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{optimize_hyper_params}
\alias{optimize_hyper_params}
\title{Optimize the hyper parameters of a GP}
\usage{
optimize_hyper_params(X, y, kernel.type = rbf, order = 5/2, noise.var = 1,
  verbose = 0)
}
\arguments{
\item{X}{An n x d matrix of covariates for observed data}

\item{y}{An n dimensional vector of outputs for observed data}

\item{order}{The order of the kernel, defaults to 5/2}

\item{noise.var}{The variance of the noise around the function}

\item{verbose}{Level of information printed out, defaults to 0}

\item{kernel}{The covariance kernel of the process, defaults to rbf}
}
\value{
Hyper parameters which optimize the log marginal likelihood
}
\description{
Optimize the hyper parameters of a GP
}

