% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{pdiff}
\alias{pdiff}
\title{Compute the pairwise differences between X and Y}
\usage{
pdiff(X, Y)
}
\arguments{
\item{X}{An nx x d matrix}

\item{Y}{An ny x d matrix}
}
\value{
An nx * ny / 2 x d matrix of the pairwise distances in all dimensions
}
\description{
Compute the pairwise differences between X and Y
}

