% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{pdist_scaled}
\alias{pdist_scaled}
\title{Compute the scaled pairwise distances between X and Y}
\usage{
pdist_scaled(X, Y, scales)
}
\arguments{
\item{X}{An nx x d matrix}

\item{Y}{An ny x d matrix}

\item{scales}{The scale parameters}
}
\value{
An nx x ny matrix of scaled distances
}
\description{
Compute the scaled pairwise distances between X and Y
}

