% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{plot.GaussianProcess}
\alias{plot.GaussianProcess}
\title{Plot a Gaussian Process}
\usage{
\method{plot}{GaussianProcess}(gp.obj, xlim = NULL, num.points = 1000, ...)
}
\arguments{
\item{gp.obj}{A gaussianProcess object}

\item{xlim}{The limits of the x-axis in the plot. Defaults to the limits of the training data plus one-quarter of the range at either end.}

\item{num.points}{The number of data points to calculate for the plot}

\item{...}{Additional parameters to pass to \code{plot}}
}
\description{
Plots the posterior mean and first three standard deviations of a one-dimensional Gaussian process.
}

