% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic.R
\name{posterior.laplace.approximation}
\alias{posterior.laplace.approximation}
\title{Calculate Laplace Approximation of a GP}
\usage{
posterior.laplace.approximation(gp.obj)
}
\arguments{
\item{gp.obj}{A trained gaussianProcess object}
}
\value{
The approximate posterior log likelihood for the input GP object.
}
\description{
Calculate the Laplace Approximation of the posterior log likelihood of a gp
}
\seealso{
\code{\link{bic.model.search}} and \code{\link{gaussianProcess}}
}

