% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{predict.GaussianProcess}
\alias{predict.GaussianProcess}
\title{Create Predictions Using a Gaussian Process}
\usage{
\method{predict}{GaussianProcess}(gp.obj, data)
}
\arguments{
\item{gp.obj}{A trained gaussianProcess object}

\item{data}{Data to predict}
}
\value{
a list containing named elements:
\itemize{
  \item \code{mean} - the predicted mean value for each data point in \code{data}.
  \item \code{var} - the variance about the predicted mean for each data point in \code{data}.
}
}
\description{
Returns the posterior mean and variance of a set of data points under a given Gaussian process
}
\examples{
x <- rnorm(50)
y <- sin(1/(x^2 + 0.15))
mt <- create.model.tree.builtin()
mt <- insert.kernel.instance(mt, 1, "SE", NULL, hyper.params=c(l=1))
gp <- create.gaussian.process(x, y, mt)
gp$fit.hyperparams(NA)

x1 <- rnorm(50)
y1.predicted <- predict(gp, x1)$mean

}

