% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{random.forest.partition.function.generator}
\alias{random.forest.partition.function.generator}
\title{Random forest kernel}
\usage{
random.forest.partition.function.generator(rf)
}
\arguments{
\item{rf}{a randomForest object}
}
\value{
a function
}
\description{
Random forest kernel
}
\seealso{
\code{\link[randomForest]{randomForest}}
}

