% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{random.partition.kernel}
\alias{random.partition.kernel}
\title{Random Partition Kernel}
\usage{
random.partition.kernel(a, b, hyper.params = NULL, additional.params)
}
\arguments{
\item{a}{first data point}

\item{b}{second data point}

\item{hyper.params}{an empy vector (included for compatibility with kernel function format)}

\item{additional.params}{A list with a named element "partitionFunction" containing the partition function which defines the kernel}
}
\value{
The covariance between \code{a} and \code{b}
}
\description{
Takes a partition function and returns the covariance between two points
derived from that partition function. The partition function should return
a vector of integers, with each integer representing the result of a different
classifier.
}

