% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperOpt.R
\name{rbf_deriv}
\alias{rbf_deriv}
\title{Compute the jth partial derivative  of the rbf kernel}
\usage{
rbf_deriv(j, hyper.params, diff.mat, K, noise.var)
}
\arguments{
\item{j}{The coordinate of the derivative to compute}

\item{hyper.params}{Hyper parameters as a vector (for rbf c(amplitude, scales))}

\item{diff.mat}{The result of \code{pdiff(X,X)} for input data \code{X}}

\item{K}{The kernel matrix of \code{X}}

\item{noise.var}{The variance of the noise around the function}
}
\value{
The jth partial derivative at \code{hyper.params}
}
\description{
Compute the jth partial derivative  of the rbf kernel
}

