% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rmse}
\alias{coef_of_determination}
\alias{coef_of_determination_boot}
\alias{mae}
\alias{mae_boot}
\alias{regression_summary_stats}
\alias{regression_summary_stats_boot}
\alias{rmse}
\alias{rmse_boot}
\title{Regression fit metrics}
\usage{
rmse(prediction, actual)

rmse_boot(data, indices, actual)

coef_of_determination(prediction, actual)

coef_of_determination_boot(data, indices, actual)

mae(prediction, actual)

mae_boot(data, indices, actual)

regression_summary_stats(prediction, actual)

regression_summary_stats_boot(data, indices, actual)
}
\arguments{
\item{prediction}{predicted values}

\item{actual}{actual values}

\item{data}{predicted values}

\item{indices}{indices included in bootstrap sample}
}
\value{
Statistic value (or vector of values)
}
\description{
Regression fit metrics
}

