% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{sample.functions.from.kernel}
\alias{sample.functions.from.kernel}
\title{Sample a function from a Gaussian Process Prior}
\usage{
sample.functions.from.kernel(x, kernel, sigma.n, hyper.params,
  num.functions = 1)
}
\arguments{
\item{x}{Points to sample at}

\item{kernel}{Kernel of the Gaussian Process}

\item{sigma.n}{Standard deviation of the noise}

\item{hyper.params}{Hyperparameters of the kernel}

\item{num.functions}{The number of different functions to generate}
}
\value{
A numeric vector of values at the sampled points
}
\description{
Sample a function from a Gaussian Process Prior
}

