% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{sample.functions.from.model.tree}
\alias{sample.functions.from.model.tree}
\title{Sample a function from a Gaussian Process Prior}
\usage{
sample.functions.from.model.tree(x, model.tree, sigma.n, num.functions = 1)
}
\arguments{
\item{x}{Points to sample at}

\item{model.tree}{Model tree which defines the kernel of the Gaussian Process}

\item{sigma.n}{Standard deviation of the noise}

\item{num.functions}{The number of different functions to generate}
}
\value{
A numeric vector of values at the sampled points
}
\description{
Sample a function from a Gaussian Process Prior
}

