% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictGP.R
\name{sample.gaussianProcess}
\alias{sample.gaussianProcess}
\title{Sample from the posterior distribution for f(X) (or the posterior predictive)}
\usage{
\method{sample}{gaussianProcess}(gp, X, n, pred = F)
}
\arguments{
\item{gp}{A gaussianProcess object with data, kernel, etc.}

\item{X}{An n x d matrix of new observations}

\item{n}{Number of samples}

\item{pred}{Boolean, whether to sample from the posterior predictive}
}
\value{
An n dimensional vector of samples from the posterior (predictive)
}
\description{
Sample from the posterior distribution for f(X) (or the posterior predictive)
}

