% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelUtils.R
\name{scale_kernel}
\alias{scale_kernel}
\title{Scale a kernel}
\usage{
scale_kernel(kernel, kernel_class_name)
}
\arguments{
\item{kernel}{a kernel object}

\item{kernel_class_name}{a string naming \code{kernel}}
}
\value{
a kernel object
}
\description{
Scale a kernel
}

