% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelUtils.R
\name{scale_kernel_list}
\alias{scale_kernel_list}
\title{Scale a list of kernel objects}
\usage{
scale_kernel_list(kernel_list)
}
\arguments{
\item{kernel_list}{a list of Kernel objects}
}
\value{
a list of kernel objects
}
\description{
Scale a list of kernel objects
}

