% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{summary.ModelTree}
\alias{summary.ModelTree}
\title{Print a Summary of a Model Tree}
\usage{
\method{summary}{ModelTree}(model.tree)
}
\arguments{
\item{model.tree}{ModelTree object to summarise}
}
\description{
Print a Summary of a Model Tree
}

