% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{test.kernel.grad}
\alias{test.kernel.grad}
\title{Test an Analytic Gradient Against a Numeric Gradient}
\usage{
test.kernel.grad(k, k.grad, hyper.param.names, additional.params,
  repetitions = 1000)
}
\arguments{
\item{k}{a kernel function}

\item{k.grad}{the analytic gradient to test}

\item{hyper.param.names}{names of the kernel's hyperparameters}

\item{additional.params}{any additional parameters of the kernel}

\item{repetitions}{number of repetitions to attempt}
}
\value{
If the gradients match, returns TRUE. If a mismatch is found, returns a list with named entries:
\itemize{
  \item a, b, hyper.params - the parameters passed to the gradient functions
  \item grad - return value of the analytic gradient
  \item grad.num - return value of the numerical gradient
  \item max.diff - the maximum relative difference observed between the two gradient functions so far (including the failed set of inputs)
  \item i - the number of inputs tested (including the failed set of inputs)
}
}
\description{
Creates a numerical approximation to the gradient function of a kernel (using \code{\link{create.numeric.grad}})
and compares it to an analytic gradient function. If the maximum relative difference exceeds 0.01, returns the failing inputs,
otherwise returns TRUE.
}

