% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structureRecreation.R
\name{test.structure.recreation}
\alias{test.structure.recreation}
\title{Test the ability of the BIC search to identify the data structure}
\usage{
test.structure.recreation(base.generator.model, base.search.model,
  num.expansions = 1, hyper.param.range = c(0, 10), sigma.n.range = c(0.1,
  1), xlim = c(-10, 10), function.points = 1000, max.points.sampled = 50,
  min.points.sampled = 50, sample.size.ratio = 2, max.new.nodes = 2)
}
\description{
Test the ability of the BIC search to identify the data structure
}

