% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelTree.R
\name{clone.env}
\alias{clone.env}
\title{Clone an environment}
\usage{
clone.env(orig.env)
}
\arguments{
\item{orig.env}{the environment to clone}
}
\value{
a new environment whose contents match \code{orig.env}
}
\description{
Clone an environment
}
\examples{
mt <- create.model.tree.builtin()
mt2 <- clone.env(mt)
}

