% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covMatUtils.R
\name{get.inst.cov.mats}
\alias{get.inst.cov.mats}
\title{Get Kernel Instance Covariance Matrices, Grads or Hessians}
\usage{
get.inst.cov.mats(x, model.tree, return.type = c("cov", "grad", "hess"),
  cache = NULL)
}
\arguments{
\item{x}{A numeric matrix}

\item{model.tree}{A modelTree object}

\item{return.type}{Whether to return the covariance matrices, the associated grads, or the hessians}
}
\value{
A named list containing covariance matrices (or grad arrays), with one entry for each kernel instance.
}
\description{
Calculates the covariance matrices (or associated grads) for the kernel instances appearing in a model tree
}

