% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGP.R
\name{plot.gaussianProcess}
\alias{plot.gaussianProcess}
\title{Plot the predictive distribution of new data. Only works with 1 dimension}
\usage{
\method{plot}{gaussianProcess}(gp, Xnew, int.size = 1.96, y.var = FALSE,
  plot.points = TRUE)
}
\arguments{
\item{gp}{gaussianProcess object}

\item{int.size}{Posterior intervals are mean +- int.size * var, defaults to 1.96}

\item{y.var}{Whether to show predictive distribution for new y, defaults to False}

\item{plot.points}{Whether to plot the data as well, defaults to False}

\item{xnew}{Test data points to plot}
}
\description{
Plot the predictive distribution of new data. Only works with 1 dimension
}

