% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernels.R
\name{rbf}
\alias{rbf}
\title{Compute the RBF (Squared Exponential) kernel between two matrices}
\usage{
rbf(X, Y, scale = 1, amplitude = 1)
}
\arguments{
\item{X}{An n x d matrix}

\item{Y}{An n x d matrix}

\item{scale}{The scale parameter, defaults to 1}

\item{amplitude}{A parameter which is multiplied by the kernel, defaults to 1}
}
\value{
An n x n kernel matrix formed from \code{X} and \code{Y}
}
\description{
Compute the RBF (Squared Exponential) kernel between two matrices
}

